+++ benches/regex_bench.rs
@@ -2,22 +2,21 @@ use std::str::FromStr;
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use once_cell::sync::Lazy;
 use serde_json::{json, Value};
-use jsonpath_rust::{CONFIG, JsonPathFinder, JsonPathInst, JsonPathQuery};
+use jsonpath_rust::{JsonPathFinder, JsonPathInst, JsonPathQuery};
 use jsonpath_rust::path::config::cache::{DefaultRegexCacheInst, RegexCache};
 use jsonpath_rust::path::config::JsonPathConfig;
 
 
-
-fn regex_perf_test_after() {
+fn regex_perf_test_with_cache(cfg: JsonPathConfig) {
     let json = Box::new(json!({
             "author":"abcd(Rees)",
         }));
 
-    let _v = (json, CONFIG.clone()).path("$.[?(@.author ~= '.*(?i)d\\(Rees\\)')]")
+    let _v = (json, cfg).path("$.[?(@.author ~= '.*(?i)d\\(Rees\\)')]")
         .expect("the path is correct");
 }
 
-fn regex_perf_test_before() {
+fn regex_perf_test_without_cache() {
     let json = Box::new(json!({
             "author":"abcd(Rees)",
         }));
@@ -27,9 +26,10 @@ fn regex_perf_test_before() {
 }
 
 pub fn criterion_benchmark(c: &mut Criterion) {
-    c.bench_function("regex bench before", |b| b.iter(|| regex_perf_test_before()));
-    c.bench_function("regex bench after", |b| {
-        b.iter(|| regex_perf_test_after())
+    let cfg = JsonPathConfig::new(RegexCache::Implemented(DefaultRegexCacheInst::default()));
+    c.bench_function("regex bench without cache", |b| b.iter(|| regex_perf_test_without_cache()));
+    c.bench_function("regex bench with cache", |b| {
+        b.iter(|| regex_perf_test_with_cache(cfg.clone()))
     });
 }