+++ src/path/config.rs
@@ -2,8 +2,10 @@ pub mod cache;
 
 use crate::path::config::cache::{RegexCache};
 
+/// Configuration to adjust the jsonpath search
 #[derive(Clone, Default)]
 pub struct JsonPathConfig {
+    /// cache to provide
     pub regex_cache: RegexCache,
 }