+++ src/path/top.rs
@@ -4,6 +4,7 @@ use crate::JsonPathValue::{NewValue, NoValue, Slice};
 use crate::{jsp_idx, jsp_obj, JsPathStr};
 use serde_json::value::Value::{Array, Object};
 use serde_json::{json, Value};
+use crate::path::config::JsonPathConfig;
 
 /// to process the element [*]
 pub(crate) struct Wildcard {}
@@ -103,7 +104,7 @@ impl<'a> Path<'a> for FnPath {
     fn flat_find(
         &self,
         input: Vec<JsonPathValue<'a, Self::Data>>,
-        is_search_length: bool
+        is_search_length: bool,
     ) -> Vec<JsonPathValue<'a, Self::Data>> {
         // todo rewrite
         if JsonPathValue::only_no_value(&input) {
@@ -156,6 +157,7 @@ impl<'a> Path<'a> for ObjectField<'a> {
         vec![res]
     }
 }
+
 /// the top method of the processing ..*
 pub(crate) struct DescentWildcard;
 
@@ -259,7 +261,7 @@ impl<'a> Chain<'a> {
             is_search_length,
         }
     }
-    pub fn from(chain: &'a [JsonPath], root: &'a Value) -> Self {
+    pub fn from(chain: &'a [JsonPath], root: &'a Value, cfg: JsonPathConfig) -> Self {
         let chain_len = chain.len();
         let is_search_length = if chain_len > 2 {
             let mut res = false;
@@ -299,7 +301,7 @@ impl<'a> Chain<'a> {
         };
 
         Chain::new(
-            chain.iter().map(|p| json_path_instance(p, root)).collect(),
+            chain.iter().map(|p| json_path_instance(p, root, cfg.clone())).collect(),
             is_search_length,
         )
     }
@@ -368,17 +370,17 @@ mod tests {
         let field4 = path!("array");
         let field5 = path!("object");
 
-        let path_inst = json_path_instance(&path!($), &json);
+        let path_inst = json_path_instance(&path!($), &json,Default::default());
         assert_eq!(path_inst.find(jp_v!(&json)), jp_v!(&json;"$",));
 
-        let path_inst = json_path_instance(&field1, &json);
+        let path_inst = json_path_instance(&field1, &json,Default::default());
         let exp_json =
             json!({"k":{"f":42,"array":[0,1,2,3,4,5],"object":{"field1":"val1","field2":"val2"}}});
         assert_eq!(path_inst.find(jp_v!(&json)), jp_v!(&exp_json;".['v']",));
 
         let chain = chain!(path!($), field1.clone(), field2.clone(), field3);
 
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
         let exp_json = json!(42);
         assert_eq!(
             path_inst.find(jp_v!(&json)),
@@ -392,7 +394,7 @@ mod tests {
             field4.clone(),
             path!(idx!(3))
         );
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
         let exp_json = json!(3);
         assert_eq!(
             path_inst.find(jp_v!(&json)),
@@ -407,7 +409,7 @@ mod tests {
             field4.clone(),
             path!(index)
         );
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
         let one = json!(1);
         let tree = json!(3);
         assert_eq!(
@@ -423,7 +425,7 @@ mod tests {
             field4,
             path!(union)
         );
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
         let tree = json!(1);
         let two = json!(2);
         assert_eq!(
@@ -433,7 +435,7 @@ mod tests {
 
         let union = idx!("field1", "field2");
         let chain = chain!(path!($), field1.clone(), field2, field5, path!(union));
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
         let one = json!("val1");
         let two = json!("val2");
         assert_eq!(
@@ -443,16 +445,18 @@ mod tests {
             &two;"$.['v'].['k'].['object'].['field2']")
         );
     }
+
     #[test]
     fn path_descent_arr_test() {
         let json = json!([{"a":1}]);
         let chain = chain!(path!($), path!(.."a"));
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
 
         let one = json!(1);
         let expected_res = jp_v!(&one;"$[0].['a']",);
         assert_eq!(path_inst.find(jp_v!(&json)), expected_res)
     }
+
     #[test]
     fn deep_path_test() {
         let value = json!([1]);
@@ -467,7 +471,7 @@ mod tests {
             "key1": [1]
         });
         let chain = chain!(path!($), path!(..*));
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
 
         let arr = json!([1]);
         let one = json!(1);
@@ -475,6 +479,7 @@ mod tests {
         let expected_res = jp_v!(&arr;"$.['key1']",&one;"$.['key1'][0]");
         assert_eq!(path_inst.find(jp_v!(&json)), expected_res)
     }
+
     #[test]
     fn path_descent_w_nested_array_test() {
         let json = json!(
@@ -482,7 +487,7 @@ mod tests {
             "key2" : [{"a":1},{}]
         });
         let chain = chain!(path!($), path!(..*));
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
 
         let arr2 = json!([{"a": 1},{}]);
         let obj = json!({"a": 1});
@@ -515,7 +520,7 @@ mod tests {
             }
         });
         let chain = chain!(path!($), path!(..*));
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
 
         let key1 = json!([1]);
         let one = json!(1);
@@ -552,6 +557,7 @@ mod tests {
         ];
         assert_eq!(path_inst.find(jp_v!(&json)), expected_res)
     }
+
     #[test]
     fn path_descent_test() {
         let json = json!(
@@ -568,7 +574,7 @@ mod tests {
             }
         });
         let chain = chain!(path!($), path!(.."key1"));
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
 
         let res1 = json!([1, 2, 3]);
         let res2 = json!("key1");
@@ -593,7 +599,7 @@ mod tests {
         });
 
         let chain = chain!(path!($), path!(*));
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
 
         let res1 = json!([1, 2, 3]);
         let res2 = json!("key");
@@ -612,7 +618,7 @@ mod tests {
         });
 
         let chain = chain!(path!($), path!(*), function!(length));
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
 
         assert_eq!(
             path_inst.flat_find(vec![jp_v!(&json)], true),
@@ -620,7 +626,7 @@ mod tests {
         );
 
         let chain = chain!(path!($), path!("key1"), function!(length));
-        let path_inst = json_path_instance(&chain, &json);
+        let path_inst = json_path_instance(&chain, &json,Default::default());
         assert_eq!(
             path_inst.flat_find(vec![jp_v!(&json)], false),
             vec![jp_v!(json!(3))]